Add scaling factor to JTC
This adds a scaling factor between 0 and 1 to the JTC so that the trajectory
time inside the controller is extended respectively. A value of 0.5 means
that trajectory execution will take twice as long as the trajectory states.

The scaling factor itself is read from the hardware for now.

Fixed formatting

use reset+initRT due to missing writeFromRT

Co-authored-by: Christoph Fröhlich <christophfroehlich@users.noreply.github.com>

synchronization of scaling factor with hw optional, add service for setting
of scaling factor